backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
openai==1.3.0
azure-search-documents==11.4.0
azure-core==1.29.5
pydantic==2.5.0
python-multipart==0.0.6
httpx==0.25.2
sse-starlette==1.8.1 
pypdf
python-docx
tiktoken
tenacity
nltk 